Added wordpress_format.py for citations
wordpress_format.py formats the citations for the SOCAT website from the DOI. This script will fetch the metadata needed from crossref.org and will automatically copy the citation string. If the article is not yet published in a physical journal, some of the information, such as issue number and page number might be missing. Created by MarenKK